requirements.txt
streamlit
nbstripout
matplotlib
sqlalchemy
polars
openai
matplotlib
xlsxwriter
openpyxl
black
cryptography
python-dotenv
plotly